optimized get_next_line (it's now 2x as fast!)

+++ get_next_line/libft/libft.h
@@ -77,7 +77,7 @@ int					ft_strequ(char const *s1, char const *s2);
 void				ft_striter(char *s, void (*f)(char *));
 void				ft_striteri(char *s, void (*f)(unsigned int, char *));
 char				*ft_strjoin(char const *s1, char const *s2);
-unsigned long		ft_strlcat(char *dst, const char *src, unsigned long size);
+size_t				ft_strlcat(char *dst, const char *src, unsigned long size);
 size_t				ft_strlen(const char *s);
 char				*ft_strmap(char const *s, char (*f)(char));
 char				*ft_strmapi(char const *s, char (*f)(unsigned int, char));
